chore: move fetch objs into their own file
